Enable serverextension in CI! Add publish integration tests.
- Update README
- also, ignore node_modules

Right now we're not automatically configuring the jupyter server. This 
manifests as a situation where requests are sent to the endpoint 
expecting a json payload and it gets back the 404 html page (which it 
doesn't know how to process). The real solution to this is to 
autoinstall the extension, but that won't help in the development case 
so it needs to be documented.